requirements/tareas.txt
Django==5.1.7
djangorestframework
gunicorn
psycopg2-binary
django-cors-headers
requests